Refactor the over_sampling
