+++ imblearn/base.py
@@ -15,7 +15,7 @@ from sklearn.externals import six
 from sklearn.utils import check_X_y
 from sklearn.utils.validation import check_is_fitted
 
-from .utils import hash_X_y
+from .utils import check_ratio, check_target_type, hash_X_y
 
 
 class SamplerMixin(six.with_metaclass(ABCMeta, BaseEstimator)):
@@ -156,3 +156,42 @@ class SamplerMixin(six.with_metaclass(ABCMeta, BaseEstimator)):
         logger = logging.getLogger(__name__)
         self.__dict__.update(dict)
         self.logger = logger
+
+
+class BaseSampler(SamplerMixin):
+    """Base class for sampling algorithms.
+
+    Warning: This class should not be used directly. Use the derive classes
+    instead.
+    """
+
+    def __init__(self, ratio='auto', random_state=None, sampling_type=None):
+        self.ratio = ratio
+        self.random_state = random_state
+        self.sampling_type = sampling_type
+        self.logger = logging.getLogger(__name__)
+
+    def fit(self, X, y):
+        """Find the classes statistics before to perform sampling.
+
+        Parameters
+        ----------
+        X : ndarray, shape (n_samples, n_features)
+            Matrix containing the data which have to be sampled.
+
+        y : ndarray, shape (n_samples, )
+            Corresponding label for each sample in X.
+
+        Returns
+        -------
+        self : object,
+            Return self.
+
+        """
+        X, y = check_X_y(X, y)
+        y = check_target_type(y)
+        self.X_hash_, self.y_hash_ = hash_X_y(X, y)
+        # self.sampling_type is already checked in check_ratio
+        self.ratio_ = check_ratio(self.ratio, y, self.sampling_type)
+
+        return self
