+++ imblearn/under_sampling/base.py
@@ -4,15 +4,10 @@ Base class for the under-sampling method.
 # Authors: Guillaume Lemaitre <g.lemaitre58@gmail.com>
 # License: MIT
 
-import logging
-
-from sklearn.utils import check_X_y
-
-from ..base import SamplerMixin
-from ..utils import check_ratio, check_target_type, hash_X_y
+from ..base import BaseSampler
 
 
-class BaseUnderSampler(SamplerMixin):
+class BaseUnderSampler(BaseSampler):
     """Base class for under-sampling algorithms.
 
     Warning: This class should not be used directly. Use the derive classes
@@ -20,36 +15,12 @@ class BaseUnderSampler(SamplerMixin):
     """
 
     def __init__(self, ratio='auto', random_state=None):
-        self.ratio = ratio
-        self.random_state = random_state
-        self.logger = logging.getLogger(__name__)
-
-    def fit(self, X, y):
-        """Find the classes statistics before to perform sampling.
-
-        Parameters
-        ----------
-        X : ndarray, shape (n_samples, n_features)
-            Matrix containing the data which have to be sampled.
-
-        y : ndarray, shape (n_samples, )
-            Corresponding label for each sample in X.
-
-        Returns
-        -------
-        self : object,
-            Return self.
-
-        """
-        X, y = check_X_y(X, y)
-        y = check_target_type(y)
-        self.X_hash_, self.y_hash_ = hash_X_y(X, y)
-        self.ratio_ = check_ratio(self.ratio, y, 'under-sampling')
-
-        return self
+        super(BaseUnderSampler, self).__init__(ratio=ratio,
+                                               random_state=random_state,
+                                               sampling_type='under-sampling')
 
 
-class BaseCleaningSampler(SamplerMixin):
+class BaseCleaningSampler(BaseSampler):
     """Base class for under-sampling algorithms.
 
     Warning: This class should not be used directly. Use the derive classes
@@ -57,30 +28,6 @@ class BaseCleaningSampler(SamplerMixin):
     """
 
     def __init__(self, ratio='auto', random_state=None):
-        self.ratio = ratio
-        self.random_state = random_state
-        self.logger = logging.getLogger(__name__)
-
-    def fit(self, X, y):
-        """Find the classes statistics before to perform sampling.
-
-        Parameters
-        ----------
-        X : ndarray, shape (n_samples, n_features)
-            Matrix containing the data which have to be sampled.
-
-        y : ndarray, shape (n_samples, )
-            Corresponding label for each sample in X.
-
-        Returns
-        -------
-        self : object,
-            Return self.
-
-        """
-        X, y = check_X_y(X, y)
-        y = check_target_type(y)
-        self.X_hash_, self.y_hash_ = hash_X_y(X, y)
-        self.ratio_ = check_ratio(self.ratio, y, 'clean-sampling')
-
-        return self
+        super(BaseCleaningSampler, self).__init__(ratio=ratio,
+                                                  random_state=random_state,
+                                                  sampling_type='clean-sampling')
